# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pokémon Knowledge Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pokemon:
    get:
      operationId: listPokemon
      tags: [pokemon]
      summary: List all Pokémon with optional filters
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 24 }
        - name: generation
          in: query
          schema: { type: integer }
        - name: type
          in: query
          schema: { type: string }
        - name: legendary
          in: query
          schema: { type: boolean }
        - name: mythical
          in: query
          schema: { type: boolean }
        - name: sortBy
          in: query
          schema: { type: string, enum: [id, name, baseStatTotal, height, weight] }
        - name: sortOrder
          in: query
          schema: { type: string, enum: [asc, desc] }
      responses:
        "200":
          description: Paginated Pokémon list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PokemonListResponse"
  /pokemon/{id}:
    get:
      operationId: getPokemon
      tags: [pokemon]
      summary: Get full Pokémon details by ID or name
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Full Pokémon details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PokemonDetail"
        "404":
          description: Not found
  /pokemon/{id}/evolution-chain:
    get:
      operationId: getPokemonEvolutionChain
      tags: [pokemon]
      summary: Get evolution chain for a Pokémon
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Evolution chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvolutionChain"
  /pokemon/{id}/forms:
    get:
      operationId: getPokemonForms
      tags: [pokemon]
      summary: Get all forms (regional, mega, gigantamax, alternate)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pokémon forms list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PokemonForm"
  /pokemon/{id}/moves:
    get:
      operationId: getPokemonMoves
      tags: [pokemon]
      summary: Get moves learned by a Pokémon
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Move list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PokemonMove"
  /pokemon/{id}/cards:
    get:
      operationId: getPokemonCards
      tags: [pokemon]
      summary: Get trading cards featuring a Pokémon
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Trading cards list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradingCard"
  /pokemon/compare:
    get:
      operationId: comparePokemon
      tags: [pokemon]
      summary: Compare multiple Pokémon stats
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: array
            items: { type: string }
      responses:
        "200":
          description: Comparison data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PokemonDetail"

  # ─── Types ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    PokemonSummary:
      type: object
      required: [id, nationalDexNumber, name, generation, types, spriteUrl, artworkUrl, baseStatTotal, isLegendary, isMythical]
      properties:
        id: { type: integer }
        nationalDexNumber: { type: integer }
        name: { type: string }
        generation: { type: integer }
        types:
          type: array
          items: { type: string }
        spriteUrl: { type: string }
        artworkUrl: { type: string }
        baseStatTotal: { type: integer }
        isLegendary: { type: boolean }
        isMythical: { type: boolean }
        isParadox: { type: boolean }
        isUltraBeast: { type: boolean }
        color: { type: string }
        isFavorited: { type: boolean }
    PokemonDetail:
      type: object
      required: [id, nationalDexNumber, name, generation, types, stats, spriteUrl, artworkUrl]
      properties:
        id: { type: integer }
        nationalDexNumber: { type: integer }
        name: { type: string }
        category: { type: string }
        description: { type: string }
        generation: { type: integer }
        height: { type: number }
        weight: { type: number }
        color: { type: string }
        shape: { type: string }
        habitat: { type: ["string", "null"] }
        genderRatio: { type: ["number", "null"] }
        captureRate: { type: integer }
        baseFriendship: { type: integer }
        growthRate: { type: string }
        isLegendary: { type: boolean }
        isMythical: { type: boolean }
        isParadox: { type: boolean }
        isUltraBeast: { type: boolean }
        types:
          type: array
          items: { type: string }
        stats:
          $ref: "#/components/schemas/PokemonStats"
        abilities:
          type: array
          items:
            $ref: "#/components/schemas/PokemonAbility"
        spriteUrl: { type: string }
        artworkUrl: { type: string }
        shinySpriteUrl: { type: ["string", "null"] }
        shinyArtworkUrl: { type: ["string", "null"] }
        typeEffectiveness:
          $ref: "#/components/schemas/TypeEffectiveness"
        eggGroups:
          type: array
          items: { type: string }
        isFavorited: { type: boolean }
    PokemonStats:
      type: object
      required: [hp, attack, defense, specialAttack, specialDefense, speed, total]
      properties:
        hp: { type: integer }
        attack: { type: integer }
        defense: { type: integer }
        specialAttack: { type: integer }
        specialDefense: { type: integer }
        speed: { type: integer }
        total: { type: integer }
    PokemonAbility:
      type: object
      required: [name, isHidden]
      properties:
        name: { type: string }
        description: { type: string }
        isHidden: { type: boolean }
    TypeEffectiveness:
      type: object
      properties:
        doubleDamageTo:
          type: array
          items: { type: string }
        halfDamageTo:
          type: array
          items: { type: string }
        noDamageTo:
          type: array
          items: { type: string }
        doubleDamageFrom:
          type: array
          items: { type: string }
        halfDamageFrom:
          type: array
          items: { type: string }
        noDamageFrom:
          type: array
          items: { type: string }
    PokemonListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PokemonSummary"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ── Evolution ───────────────────────────────────────────────────────────
    EvolutionChain:
      type: object
      required: [chain]
      properties:
        chain:
          $ref: "#/components/schemas/EvolutionNode"
    EvolutionNode:
      type: object
      required: [pokemon]
      properties:
        pokemon:
          $ref: "#/components/schemas/PokemonSummary"
        evolvesTo:
          type: array
          items:
            $ref: "#/components/schemas/EvolutionNode"
        evolutionDetails:
          $ref: "#/components/schemas/EvolutionDetails"
    EvolutionDetails:
      type: object
      properties:
        trigger: { type: string }
        minLevel: { type: ["integer", "null"] }
        item: { type: ["string", "null"] }
        friendship: { type: ["integer", "null"] }
        timeOfDay: { type: ["string", "null"] }
        heldItem: { type: ["string", "null"] }
        specialRequirement: { type: ["string", "null"] }

    # ── Forms ───────────────────────────────────────────────────────────────
    PokemonForm:
      type: object
      required: [id, name, formType, spriteUrl, artworkUrl]
      properties:
        id: { type: integer }
        name: { type: string }
        formType:
          type: string
          enum: [normal, shiny, regional, mega, gigantamax, alternate]
        region: { type: ["string", "null"] }
        spriteUrl: { type: string }
        artworkUrl: { type: string }
        types:
          type: array
          items: { type: string }
        stats:
          $ref: "#/components/schemas/PokemonStats"

    # ── Moves ───────────────────────────────────────────────────────────────
    Move:
      type: object
      required: [id, name, type, category]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        category:
          type: string
          enum: [Physical, Special, Status]
        power: { type: ["integer", "null"] }
        accuracy: { type: ["integer", "null"] }
        pp: { type: integer }
        priority: { type: integer }
        description: { type: string }
    PokemonMove:
      type: object
      required: [move, learnMethod]
      properties:
        move:
          $ref: "#/components/schemas/Move"
        learnMethod:
          type: string
          enum: [level-up, tm, egg, tutor]
        levelLearnedAt: { type: ["integer", "null"] }

    # ── Types ───────────────────────────────────────────────────────────────
...
    TradingCard:
      type: object
      required: [id, name, set, cardNumber, rarity, imageUrl]
      properties:
        id: { type: integer }
        name: { type: string }
        set: { type: string }
        cardNumber: { type: string }
        hp: { type: ["integer", "null"] }
        rarity: { type: string }
        illustrator: { type: ["string", "null"] }
        imageUrl: { type: string }

    # ── Search ──────────────────────────────────────────────────────────────
...