# lib/api-zod/src/generated/types/evolutionDetails.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pokémon Knowledge Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EvolutionDetails {
  trigger?: string;
  /** @nullable */
  minLevel?: number | null;
  /** @nullable */
  item?: string | null;
  /** @nullable */
  friendship?: number | null;
  /** @nullable */
  timeOfDay?: string | null;
  /** @nullable */
  heldItem?: string | null;
  /** @nullable */
  specialRequirement?: string | null;
}