# lib/api-zod/src/generated/types/evolutionNode.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pokémon Knowledge Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { EvolutionDetails } from './evolutionDetails';
import type { PokemonSummary } from './pokemonSummary';

export interface EvolutionNode {
  pokemon: PokemonSummary;
  evolvesTo?: EvolutionNode[];
  evolutionDetails?: EvolutionDetails;
}