# lib/api-zod/src/generated/types/pokemonAbility.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pokémon Knowledge Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PokemonAbility {
  name: string;
  description?: string;
  isHidden: boolean;
}