# lib/api-zod/src/generated/types/pokemonSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Pokémon Knowledge Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PokemonSummary {
  id: number;
  nationalDexNumber: number;
  name: string;
  generation: number;
  types: string[];
  spriteUrl: string;
  artworkUrl: string;
  baseStatTotal: number;
  isLegendary: boolean;
  isMythical: boolean;
  isParadox?: boolean;
  isUltraBeast?: boolean;
  color?: string;
  isFavorited?: boolean;
}